[Feat] #41 routine detail bool type
[Feat] #41 routine detail bool type
